introduce save_data feature for ryftone search engine

+++ search/ryftone/ryftone.go
@@ -98,36 +98,51 @@ func (engine *Engine) process(task *Task, cfg *search.Config, res *search.Result
 	}
 
 	// INDEX results file
-	var indexFile string
 	if len(task.IndexFileName) != 0 {
-		// file path relative to `ryftone` mountpoint (including just instance)
-		indexFile = filepath.Join(engine.Instance, task.IndexFileName)
+		if len(cfg.KeepIndexAs) != 0 {
+			task.IndexFileName = cfg.KeepIndexAs
+			task.KeepIndexFile = true
+			if !strings.HasSuffix(task.IndexFileName, ".txt") {
+				// ryft adds .txt anyway, so if this extension is missed
+				// other code won't properly work!
+				task.IndexFileName += ".txt"
+				task.log().WithField("index", task.IndexFileName).
+					Warnf("[%s]: index file name was updated to have TXT extension", TAG)
+			}
+		} else {
+			// file path relative to `ryftone` mountpoint (including just instance)
+			task.IndexFileName = filepath.Join(engine.Instance, task.IndexFileName)
+		}
 	}
 
 	// DATA results file
-	var dataFile string
 	if len(task.DataFileName) != 0 {
-		// file path relative to `ryftone` mountpoint (including just instance)
-		dataFile = filepath.Join(engine.Instance, task.DataFileName)
+		if len(cfg.KeepDataAs) != 0 {
+			task.DataFileName = cfg.KeepDataAs
+			task.KeepDataFile = true
+		} else {
+			// file path relative to `ryftone` mountpoint (including just instance)
+			task.DataFileName = filepath.Join(engine.Instance, task.DataFileName)
+		}
 	}
 
 	// select search mode (fuzzy-hamming search by default)
 	switch cfg.Mode {
 	case "exact_search", "exact", "es":
 		err = task.dataSet.SearchExact(engine.prepareQuery(cfg.Query),
-			dataFile, indexFile, cfg.Surrounding, cfg.CaseSensitive)
+			task.DataFileName, task.IndexFileName, cfg.Surrounding, cfg.CaseSensitive)
 	case "fuzzy_hamming_search", "fuzzy_hamming", "fhs", "":
 		err = task.dataSet.SearchFuzzyHamming(engine.prepareQuery(cfg.Query),
-			dataFile, indexFile, cfg.Surrounding, cfg.Fuzziness, cfg.CaseSensitive)
+			task.DataFileName, task.IndexFileName, cfg.Surrounding, cfg.Fuzziness, cfg.CaseSensitive)
 	case "fuzzy_edit_distance_search", "fuzzy_edit_distance", "feds":
 		err = task.dataSet.SearchFuzzyEditDistance(engine.prepareQuery(cfg.Query),
-			dataFile, indexFile, cfg.Surrounding, cfg.Fuzziness, cfg.CaseSensitive, true)
+			task.DataFileName, task.IndexFileName, cfg.Surrounding, cfg.Fuzziness, cfg.CaseSensitive, true)
 	case "date_search", "date", "ds":
 		err = task.dataSet.SearchDate(engine.prepareQuery(cfg.Query),
-			dataFile, indexFile, cfg.Surrounding)
+			task.DataFileName, task.IndexFileName, cfg.Surrounding)
 	case "time_search", "time", "ts":
 		err = task.dataSet.SearchTime(engine.prepareQuery(cfg.Query),
-			dataFile, indexFile, cfg.Surrounding)
+			task.DataFileName, task.IndexFileName, cfg.Surrounding)
 	default:
 		err = fmt.Errorf("%q is unknown search mode", cfg.Mode)
 	}
@@ -199,11 +214,13 @@ func (engine *Engine) finish(err error, task *Task, res *search.Result) {
 	}
 
 	// cleanup: remove INDEX&DATA files at the end of processing
-	if !engine.KeepResultFiles {
+	if !engine.KeepResultFiles && !task.KeepIndexFile {
 		if err := engine.removeFile(task.IndexFileName); err != nil {
 			task.log().WithError(err).Warnf("[%s]: failed to remove INDEX file", TAG)
 			// WARN: error actually ignored!
 		}
+	}
+	if !engine.KeepResultFiles && !task.KeepDataFile {
 		if err := engine.removeFile(task.DataFileName); err != nil {
 			task.log().WithError(err).Warnf("[%s]: failed to remove DATA file", TAG)
 			// WARN: error actually ignored!
@@ -220,7 +237,7 @@ func (engine *Engine) processIndex(task *Task, res *search.Result) {
 	task.log().Debugf("[%s]: start INDEX processing...", TAG)
 
 	// try to open INDEX file: if operation is cancelled `file` is nil
-	path := filepath.Join(engine.MountPoint, engine.Instance, task.IndexFileName)
+	path := filepath.Join(engine.MountPoint, task.IndexFileName)
 	file, err := task.openFile(path, engine.OpenFilePollTimeout, task.cancelIndexChan)
 	if err != nil {
 		task.log().WithError(err).WithField("path", path).
@@ -292,7 +309,7 @@ func (engine *Engine) processData(task *Task, res *search.Result) {
 	task.log().Debugf("[%s]: start DATA processing...", TAG)
 
 	// try to open DATA file: if operation is cancelled `file` is nil
-	path := filepath.Join(engine.MountPoint, engine.Instance, task.DataFileName)
+	path := filepath.Join(engine.MountPoint, task.DataFileName)
 	file, err := task.openFile(path, engine.OpenFilePollTimeout, task.cancelDataChan)
 	if err != nil {
 		task.log().WithError(err).WithField("path", path).
@@ -355,8 +372,7 @@ func (engine *Engine) prepareQuery(query string) string {
 // removeFile removes INDEX or DATA file.
 func (engine *Engine) removeFile(name string) error {
 	if len(name) != 0 {
-		path := filepath.Join(engine.MountPoint,
-			engine.Instance, name) // full path
+		path := filepath.Join(engine.MountPoint, name) // full path
 		err := os.RemoveAll(path)
 		if err != nil {
 			return err
